refactor(data): organize assistant question sets
