+++ frontend/src/data/assistant/questionsBeginner.js
